requirements.txt
requests
PyGithub
schedule
openai